revise dataset to be same with notebook version

+++ training/standard_callbacks.py
@@ -60,8 +60,7 @@ def create_eval_callback(eval_name: str, loader: DataLoader, verbose=False):
                 labels_size = torch.tensor(len(labels), device=get_platform().torch_device)
                 example_count += labels_size
                 
-                labels = labels.long() 
-                total_loss += model.loss_criterion(output, labels) * labels_size
+                total_loss += model.loss_criterion(output, labels.long()) * labels_size
                 total_correct += correct(labels, output)
 
         # Share the information if distributed.
